fix: merge strings helpers-golang with helpers
