+++ src/lib/strings/index.ts
@@ -1,4 +1,3 @@
 export * from "./errors";
-export * from "./helpers-golang";
 export * from "./helpers";
 export * from "./time";
