+++ src/lib/strings/time.ts
@@ -48,13 +48,3 @@ export const niceDate = (date: number | string | Date) => {
 		minute: "2-digit"
 	})}`;
 };
-
-export const getAge = (birthDate: Date): number => {
-	var today = new Date();
-	var age = today.getFullYear() - birthDate.getFullYear();
-	var m = today.getMonth() - birthDate.getMonth();
-	if (m < 0 || (m === 0 && today.getDate() < birthDate.getDate())) {
-		age--;
-	}
-	return age;
-};
